requirements.txt

fastapi==0.115.5
uvicorn[standard]==0.32.0
httpx==0.27.2
python-dotenv==1.0.1
openai==1.52.0
psycopg2-binary==2.9.10
sentence-transformers==3.2.1
pydantic==2.7.4
pytest==8.3.3
